moved exec tests to a separate file
